Break up files into submodules
Done to make it easier to see what is going on. Smaller module files
are generally better and more "pythonic"

The file layout is now more in line with most Python projects you see
in the wild, and also seperates code from other assets.

The README.md file is just to provide myself with notes on how I got
it up and running locally.
