+++ .gitignore
@@ -1,0 +1,3 @@
+map/
+*.pyc
+*.egg-info
