+++ requirements.txt
@@ -1,0 +1,2 @@
+cffi==1.10.0
+libtcod-cffi==4.0.0
