+++ setup.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+
+from setuptools import setup
+
+setup()
